refactor: extract crates/cell
Phase 1.2 of the cached-cargo-workspace split. Lifts the WASM execution
layer into its own workspace crate. The cell crate is the single biggest
consumer of the heaviest deps (wasmtime, wasmtime-wasi, wasmtime-wasi-io,
cap-std), so isolating it from the rest of ww means edits to cli/main.rs
or rpc/ no longer invalidate any of those translation units.

Files moved (12 files, plus the wit/ directory):
- src/cell/{lib(was mod),proc,streams,streams_test,swappable,capnp_test}.rs
- src/{vfs,mount,sched,epoch,loaders,image,fs_intercept}.rs
- wit/{README.md,streams.wit,deps/} (consumed by bindgen! in proc.rs)

Heavy deps moved out of ww and into crates/cell:
- wasmtime-wasi, wasmtime-wasi-io (cell-set only)
- cap-std (fs_intercept)
- arc-swap, lru (vfs)
- New deps cell needs: dirs, rand, futures (already used by moved files)

ww still keeps: wasmtime (host.rs/runtime.rs use Engine + Config),
wasmparser (rpc/mod.rs uses it for custom-section parsing).

Backward compat: src/lib.rs re-exports the cell crate as 'pub use cell;'
so external callers using ww::cell::* continue to work without churn.
Plus 'pub use cell::{Loader, Proc, ProcBuilder};' for the convenience
re-exports the bin previously offered.

Bin imports updated:
- src/{executor,launcher,runtime}.rs: crate::cell::* -> cell::*,
  crate::{vfs,sched,epoch,loaders,image}::* -> cell::{vfs,...}::*
- src/cli/main.rs: ww::vfs/sched/epoch/etc -> ww::cell::vfs/etc;
  ww::image -> ww::cell::image

cargo check --lib --bins --tests -p ww passes clean. Full test suite
compiles (cargo test --no-run), 10+ test executables built.
