+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/atom",
     "crates/glia",
     "crates/cache",
+    "crates/cell",
     "crates/ipfs",
     "crates/membrane",
     "crates/schema-id",
@@ -62,14 +63,13 @@ clap = { version = "4.5.51", features = ["derive", "env"] }
 tracing-subscriber = { version = "0.3.20", features = ["env-filter"] }
 nix = { version = "0.30.1", features = ["socket"] }
 wasmtime = { version = "40.0.0", features = ["call-hook"] }
-wasmtime-wasi = "40.0.0"
-wasmtime-wasi-io = "40.0.0"
 capnp = "0.25.3"
 capnp-rpc = "0.25.0"
 membrane = { path = "crates/membrane" }
 atom = { path = "crates/atom" }
 stem = { path = "crates/stem" }
 cache = { path = "crates/cache" }
+cell = { path = "crates/cell" }
 libp2p-stream = "0.3.0-alpha"
 blake3 = "1"
 cid = "0.11"
@@ -79,9 +79,6 @@ axum = "0.8"
 rustyline = "15"
 dirs = "5"
 indicatif = "0.17"
-cap-std = "3.4"
-arc-swap = "1"
-lru = "0.12"
 wasmparser = "0.243"
 
 [[bin]]
