+++ crates/cell/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "cell"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1"
+tokio = { version = "1", features = ["full"] }
+tracing = "0.1"
+arc-swap = "1"
+lru = "0.12"
+cid = "0.11"
+hex = "0.4"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "json"] }
+futures = "0.3"
+rand = "0.9"
+dirs = "5"
+cap-std = "3.4"
+wasmtime = { version = "40.0.0", features = ["call-hook"] }
+wasmtime-wasi = "40.0.0"
+wasmtime-wasi-io = "40.0.0"
+
+# Workspace crates
+atom = { path = "../atom" }
+stem = { path = "../stem" }
+membrane = { path = "../membrane" }
+cache = { path = "../cache" }
+ipfs = { path = "../ipfs" }
