+++ crates/cell/src/streams.rs
