+++ crates/cell/src/swappable.rs
