+++ crates/cell/wit/streams.wit
