+++ src/lib.rs
@@ -5,7 +5,7 @@
 
 // Host-only modules (not available for WASM guests)
 #[cfg(not(target_arch = "wasm32"))]
-pub mod cell;
+pub use cell;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod daemon_config;
 #[cfg(not(target_arch = "wasm32"))]
@@ -13,15 +13,9 @@ pub mod discovery;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod dispatcher;
 #[cfg(not(target_arch = "wasm32"))]
-pub mod epoch;
-#[cfg(not(target_arch = "wasm32"))]
 pub mod executor;
 #[cfg(not(target_arch = "wasm32"))]
-pub(crate) mod fs_intercept;
-#[cfg(not(target_arch = "wasm32"))]
 pub mod host;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod image;
 #[cfg(not(target_arch = "wasm32"))]
 pub use ipfs;
 #[cfg(not(target_arch = "wasm32"))]
@@ -29,21 +23,13 @@ pub mod keys;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod launcher;
 #[cfg(not(target_arch = "wasm32"))]
-pub mod loaders;
-#[cfg(not(target_arch = "wasm32"))]
 pub mod metrics;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod mount;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod ns;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod rpc;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod runtime;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod sched;
-#[cfg(not(target_arch = "wasm32"))]
-pub mod vfs;
 
 // Re-export capnp schema modules from the membrane crate so host code can
 // use `crate::system_capnp`, `crate::routing_capnp`, `crate::stem_capnp`, etc.
