+++ src/runtime.rs
@@ -19,7 +19,7 @@ use anyhow::Result;
 use tokio::sync::{mpsc, watch};
 use wasmtime::Engine;
 
-use crate::sched::EPOCH_TICK_MS;
+use cell::sched::EPOCH_TICK_MS;
 
 // ---------------------------------------------------------------------------
 // Service trait
@@ -497,7 +497,7 @@ pub struct EpochService {
     pub epoch_tx: watch::Sender<Epoch>,
     pub confirmation_depth: u64,
     pub ipfs_client: crate::ipfs::HttpClient,
-    pub cid_tree: Option<std::sync::Arc<crate::vfs::CidTree>>,
+    pub cid_tree: Option<std::sync::Arc<cell::vfs::CidTree>>,
     /// Graceful shutdown: capabilities have this long to finish before epoch advances.
     pub drain_duration: std::time::Duration,
 }
@@ -510,7 +510,7 @@ impl Service for EpochService {
         let _span = tracing::info_span!("epoch").entered();
         rt.block_on(async move {
             tokio::select! {
-                result = crate::epoch::run_epoch_pipeline(
+                result = cell::epoch::run_epoch_pipeline(
                     self.config,
                     self.epoch_tx,
                     self.confirmation_depth,
